10_lesson/requirements.txt
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.15.0
